[gh-#1129] load raw script
