--- .github/workflows/audit.yaml
@@ -1,26 +1,0 @@
-name: NPM Audit
-
-on: [push, pull_request]
-
-concurrency:
-  group: audit--${{ github.event.pull_request.number || github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  audit:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18'
-
-      - name: Install dependencies
-        run: npm install
-
-      - name: Run NPM Audit
-        run: npm audit || exit 1
