--- packages/shared/lib/services/sync/integration/mocks/nango-integrations/action.ts
@@ -1,4 +1,0 @@
-export default async function runAction(nango: any, param: any) {
-    console.log(nango);
-    return param;
-}
