--- packages/shared/lib/services/sync/integration/mocks/nango-integrations/dependency.ts
@@ -1,8 +1,0 @@
-import { createHash } from 'crypto';
-
-export default async function fetchData(nango: any) {
-    const id = createHash('md5').update('get').digest('hex');
-    console.log(id);
-
-    return await nango.get();
-}
