--- packages/shared/lib/services/sync/integration/mocks/nango-integrations/helper.ts
@@ -1,7 +1,0 @@
-export function log(str: string) {
-    console.log(str);
-}
-
-export function add(num: number) {
-    return num + 1;
-}
