--- packages/shared/lib/services/sync/integration/mocks/nango-integrations/nango.yaml
@@ -1,20 +1,0 @@
-integrations:
-  google:
-    simple:
-      type: sync
-      runs: every 12 hours
-      returns:
-        - Test
-    importer:
-      type: sync
-      runs: every 12 hours
-      returns:
-        - Test
-    dependency:
-      type: sync
-      runs: every 12 hours
-      returns:
-        - Test
-models:
-  Test:
-    id: string
