--- packages/shared/lib/services/sync/integration/mocks/nango-integrations/simple.ts
@@ -1,9 +1,0 @@
-export default async function fetchData(nango: any) {
-    const other = someOther();
-    console.log(other);
-    return await nango.get();
-}
-
-function someOther() {
-    return 'nango';
-}
