+++ packages/shared/package.json
@@ -42,13 +42,11 @@
         "ms": "^2.1.3",
         "pg": "^8.8.0",
         "posthog-node": "^2.2.3",
-        "quickjs-emscripten": "^0.23.0",
         "rimraf": "^5.0.1",
         "semver": "^7.5.4",
         "simple-oauth2": "^5.0.0",
         "ts-node": "^10.9.1",
         "uuid": "^9.0.0",
-        "vm2": "^3.9.19",
         "winston": "^3.8.2",
         "winston-daily-rotate-file": "^4.7.1"
     },
