+++ packages/worker/package.json
@@ -30,7 +30,8 @@
         "knex": "^2.4.2",
         "md5": "^2.3.0",
         "nanoid": "3.x",
-        "uuid": "^9.0.0"
+        "uuid": "^9.0.0",
+        "vm2": "^3.9.19"
     },
     "devDependencies": {
         "@tsconfig/node16": "^1.0.0",
